+++ src/vue/components/InitConfig.vue
@@ -37,8 +37,8 @@
         key: 'password',
         username: 'Qiscus Demo',
         description: 'This account is used for demo purpose only',
-        // target_email: 'fikri@qiscus.com'
         target_email: 'fikri@qiscus.com'
+        // target_email: '6281212345678@kiwari-prod.com'
       }
     },
     methods: {
@@ -50,6 +50,7 @@
         qiscus.setUser(this.email, this.key, this.username);
         qiscus.init({
           AppId: 'dragongo',
+          // AppId: 'kiwari-prod',
           plugins: ['emoji'],
           options: {
             email: this.email,